Add support for Entry Point commands (#3021)
The discord docs are a little confusing for this feature, but I think
it's basically pretty simple. Applications with activities enabled can
define a global command of type `PrimaryEntryPoint` - based on the value
of the `handler` field, this command can be treated essentially like
normal, allowing the application to respond with a followup message when
called; or, the behavior of the command can be automatically handled by
Discord.