Clean up permissions.rs (#3046)
This PR combines the `bitflags` and `generate_permission_names` macros
into a single macro, which also generates the shorthand methods. The
first commit in this PR is the actual change but with weird formatting
to make it nicer to diff, and the second commit fixes that formatting.

This should make no user-facing diff, except from fixing missing entries
in `get_permission_names`.